pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bendgen"
version = "0.1.0"
description = "Web-based bend program generator for Langmuir BendControl"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "pydantic>=2.0",
    "ezdxf>=1.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3",
    "Pillow>=10.0",
]

[project.scripts]
bendgen = "bendgen.app:main"
